% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ontology_descendants.R
\name{get_ontology_descendants}
\alias{get_ontology_descendants}
\title{Get ontology descendants}
\usage{
get_ontology_descendants(ont, terms, include_self = TRUE, ...)
}
\arguments{
\item{ont}{An ontology of class \link[simona]{ontology_DAG}.}

\item{terms}{A vector of ontology term IDs.}

\item{include_self}{For dag_offspring() and dag_ancestors(), this controls 
whether to also include the query term itself.}

\item{...}{Additional arguments passed to  \code{simona::dag_offspring}.}
}
\value{
A named list of descendants, where the names are the input terms.
}
\description{
Get the descendants of a set of ontology terms.
}
\examples{
ont <- get_ontology("hp", terms=10)
d <- get_ontology_descendants(ont, c("HP:0000001","HP:0000002"))
}
